[TARGETING] Always create dummy entry from TimeOnSite condition
This makes sure the created/update timestamps are properly set before
starting to match. Depending on the storage, a key might expire if it is
not regularly updated. E.g. Redis sets an expire of 30 minutes on
session keys. If session storage is never updated in that timespan, the
storage will expire and timeOnSite will start from 0.